NSA-backend-main/requirements.txt
psutil==6.1.1
gevent>=22.0.0
asgiref==3.6.0
boto3==1.34.148
botocore==1.34.148
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-ses==3.5.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2pip install psycopg2
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
s3transfer==0.10.2
setuptools==71.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.12.2
urllib3==1.26.15
whitenoise==6.7.0
